fix: Remove duplicate nav bar and delete Discover page
This commit addresses two issues:

1.  A duplicate navigation bar was appearing on the student and instructor dashboards. This was caused by the dashboard layout extending the base layout, which also contained a navigation bar. The fix was to wrap the header in `base.html` in a Jinja block and then override that block in `dashboard_layout.html`.

2.  The "Discover" page was causing the application to crash due to a bug. As requested by the user, the "Discover" page and all related functionality have been removed. This includes the route, template, CSS, and JavaScript files, as well as the link in the navigation bar. All redirects that previously pointed to the "Discover" page now point to the "Home" page of the feed.
