--- static/css/discover_feed.css
@@ -1,379 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600;700&display=swap');
-
-/* --- Root Variables --- */
-:root {
-    --main-bg-color: #18191A;
-    --content-bg-color: #242526;
-    --divider-color: #3A3B3C;
-    --primary-text-color: #E4E6EB;
-    --secondary-text-color: #B0B3B8;
-    --link-color: #2D88FF;
-    --border-radius-sm: 4px;
-    --border-radius-md: 8px;
-    --border-radius-lg: 12px;
-}
-
-body.discover-page {
-    background-color: var(--main-bg-color);
-    color: var(--primary-text-color);
-    font-family: 'Poppins', sans-serif;
-}
-
-/* --- Discover Page Specific Layout --- */
-.discover-page .site-header {
-    display: none;
-}
-
-.discover-container {
-    padding: 1rem 2rem;
-}
-
-.discover-top-bar {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 1rem;
-    background-color: var(--content-bg-color);
-    border-radius: var(--border-radius-lg);
-    margin-bottom: 1.5rem;
-    border: 1px solid var(--divider-color);
-}
-
-.discover-header-left {
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-}
-
-.discover-title {
-    font-size: 1.5rem;
-    font-weight: 700;
-    color: var(--primary-text-color);
-}
-
-.discover-header-right {
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-}
-
-.header-icon {
-    font-size: 1.2rem;
-    color: var(--primary-text-color);
-}
-
-.profile-pic-icon {
-    width: 32px;
-    height: 32px;
-    border-radius: 50%;
-}
-
-/* --- Stories --- */
-.stories-container {
-    display: flex;
-    gap: 1.5rem; /* Increased gap */
-    padding-bottom: 1.5rem;
-    margin-bottom: 1.5rem;
-    border-bottom: 1px solid var(--divider-color);
-    overflow-x: auto;
-}
-
-.story {
-    text-align: center;
-    cursor: pointer;
-    flex-shrink: 0;
-}
-
-.story-avatar {
-    width: 72px; /* Slightly larger */
-    height: 72px;
-    border-radius: 50%;
-    border: 3px solid var(--link-color);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    padding: 3px;
-    background-color: var(--content-bg-color);
-}
-
-.story-avatar img {
-    width: 100%;
-    height: 100%;
-    border-radius: 50%;
-    object-fit: cover;
-}
-
-.add-story .story-avatar {
-    border-radius: var(--border-radius-md); /* Square card */
-    border: 2px dashed var(--secondary-text-color);
-}
-
-.add-story .story-avatar i {
-    color: var(--secondary-text-color);
-    font-size: 1.5rem;
-}
-
-.story span {
-    font-size: 0.85rem; /* Slightly larger font */
-    margin-top: 0.5rem;
-    display: block;
-    color: var(--primary-text-color);
-}
-
-/* --- Feed --- */
-.discover-feed {
-    max-width: 680px;
-    margin: 0 auto;
-}
-
-.post-card {
-    background-color: var(--content-bg-color);
-    border-radius: var(--border-radius-lg);
-    margin-bottom: 20px;
-    border: 1px solid var(--divider-color);
-    width: 100%;
-}
-
-.post-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 1rem;
-}
-
-.post-author-info {
-    display: flex;
-    align-items: center;
-    gap: 0.75rem;
-}
-
-.post-author-info img {
-    width: 40px;
-    height: 40px;
-    border-radius: 50%;
-    border: 1px solid var(--divider-color); /* Subtle dark outline */
-}
-
-.post-author-info strong a {
-    font-weight: 700;
-    color: var(--primary-text-color);
-    text-decoration: none;
-}
-
-.post-author-info .timestamp {
-    font-size: 0.85rem;
-    color: var(--secondary-text-color);
-}
-
-.post-content {
-    padding: 0 1rem;
-    overflow: hidden;
-    transition: max-height 0.3s ease;
-    color: var(--primary-text-color);
-}
-
-.post-content.collapsed {
-    max-height: 66px; /* Approx 3 lines of Poppins */
-}
-
-.post-content.expanded {
-    max-height: none;
-}
-
-.see-more-btn {
-    background: none;
-    border: none;
-    color: var(--link-color);
-    cursor: pointer;
-    font-weight: bold;
-    padding: 0.5rem 1rem;
-}
-
-.post-media {
-    margin-top: 1rem;
-    background-color: var(--main-bg-color);
-}
-
-.post-media img, .post-media video {
-    width: 100%;
-    max-height: 500px;
-    object-fit: cover;
-    border-radius: var(--border-radius-md);
-}
-
-.post-divider {
-    border: none;
-    border-top: 1px solid var(--divider-color);
-    margin: 0.5rem 0;
-}
-
-.post-footer {
-    padding: 0 1rem 0.5rem;
-}
-
-.post-stats {
-    display: flex;
-    justify-content: space-between;
-    padding: 0.5rem 0;
-    color: var(--secondary-text-color);
-    font-size: 0.9rem;
-}
-
-.post-actions {
-    display: flex;
-    justify-content: space-around;
-    padding: 0.5rem 0;
-}
-
-.post-action {
-    background: none;
-    border: none;
-    color: var(--secondary-text-color);
-    font-weight: 600;
-    cursor: pointer;
-    padding: 0.5rem 1rem;
-    border-radius: var(--border-radius-md);
-    transition: background-color 0.2s;
-    flex-grow: 1;
-    text-align: center;
-}
-.post-action:hover {
-    background-color: #3A3B3C;
-}
-.post-action i {
-    margin-right: 0.5rem;
-}
-
-.comment-preview {
-    padding: 0.5rem 0;
-}
-
-.comment-item {
-    display: flex;
-    align-items: flex-start;
-    gap: 0.5rem;
-    margin-bottom: 0.5rem;
-}
-
-.comment-author-avatar {
-    width: 32px;
-    height: 32px;
-    border-radius: 50%;
-}
-
-.add-comment {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    padding: 0.5rem 0;
-}
-
-.comment-input {
-    width: 100%;
-    border: 1px solid var(--divider-color);
-    border-radius: 20px;
-    padding: 0.5rem 1rem;
-    background-color: #3A3B3C;
-    color: var(--primary-text-color);
-}
-.comment-input::placeholder {
-    color: var(--secondary-text-color);
-}
-
-
-/* --- Story Viewer Modal --- */
-.story-modal {
-    background-color: rgba(0,0,0,0.9);
-}
-
-.story-viewer-content {
-    width: 100%;
-    height: 100%;
-    max-width: 400px;
-    max-height: 95vh;
-    margin: auto;
-    display: flex;
-    flex-direction: column;
-    position: relative;
-    padding: 0;
-    background: black;
-    border-radius: var(--border-radius-lg);
-}
-
-.story-progress-bars {
-    display: flex;
-    gap: 4px;
-    padding: 8px;
-    position: absolute;
-    top: 0;
-    width: calc(100% - 16px);
-    z-index: 10;
-}
-
-.progress-bar-background {
-    flex-grow: 1;
-    height: 3px;
-    background-color: rgba(255, 255, 255, 0.3);
-    border-radius: 3px;
-}
-
-.progress-bar-fill {
-    height: 100%;
-    background-color: white;
-    width: 0%;
-    transition: width 0.1s linear;
-}
-
-.story-header {
-    display: flex;
-    align-items: center;
-    gap: 10px;
-    padding: 20px 16px 10px 16px;
-    position: absolute;
-    top: 10px;
-    left: 0;
-    width: calc(100% - 32px);
-    z-index: 10;
-    color: white;
-}
-
-.story-user-avatar {
-    width: 32px;
-    height: 32px;
-    border-radius: 50%;
-}
-
-.close-story-viewer {
-    margin-left: auto;
-    font-size: 2rem;
-    cursor: pointer;
-}
-
-.story-media-container {
-    flex-grow: 1;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-.story-media-container img, .story-media-container video {
-    width: 100%;
-    height: 100%;
-    object-fit: contain;
-    border-radius: var(--border-radius-lg);
-}
-
-.story-nav {
-    position: absolute;
-    top: 50%;
-    transform: translateY(-50%);
-    background: rgba(255,255,255,0.2);
-    border: none;
-    color: white;
-    width: 40px;
-    height: 40px;
-    border-radius: 50%;
-    cursor: pointer;
-    z-index: 11;
-}
-.story-nav.prev { left: 10px; }
-.story-nav.next { right: 10px; }
