+++ templates/base.html
@@ -42,6 +42,7 @@
     </div>
     {% endif %}
 
+    {% block header %}
     <header class="site-header" role="banner">
         <div class="nav-container">
             <div class="site-logo">ScholarsNovara</div>
@@ -101,6 +102,7 @@
             </nav>
         </div>
     </header>
+    {% endblock %}
 
     <main>
         {% with messages = get_flashed_messages(with_categories=true) %}
